refactor: split app into pages, containers, and components for better structure
