+++ demo/components/index.ts
@@ -1,0 +1,2 @@
+export { FileSelector } from './file-selector';
+export { LoudnessMetricItem } from './loudness-metric-item';
