+++ demo/pages/index.ts
@@ -1,0 +1,2 @@
+export { HeroPage } from './hero-page';
+export { MeterPage } from './meter-page';
